pyproject.toml
[tool.poetry]
name = "bibbi-soji"
version = "0.1.0"
description = "Collection of Bibbi cleanup and matching scripts"
authors = ["Dan Michael O. Heggø <Dan.Michael.Heggo@bibsent.no>"]
license = "MIT"
packages = [
    { include = "soji" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
pydash = "^4.7.6"
python-dotenv = "^0.13.0"
unidecode = "^1.1.1"
lxml = "^4.5.0"
tqdm = "^4.46.0"
cachecontrol = "^0.12.6"
pyodbc = "^4.0.30"
openpyxl = "^3.0.3"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.0"
prompt_toolkit = "^3.0.5"
questionary = "^1.5.2"
sqlparams = "^3.0.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
pylint = "^2.5.2"
black = "^19.10b0"
pytest = "^5.4.2"
pycodestyle = "^2.6.0"
pyflakes = "^2.2.0"

[tool.poetry.scripts]
match_persons = "soji.console.match_persons:main"
update_persons = "soji.console.update_persons:main"
validate_persons = "soji.console.validate_persons:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
